#6 Feat: header, footer 부트스트랩 사용
기존 styled-component에서 부트스트랩 변경에 따른 footer,header 수정
