+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_KAKAO_API = 280daea7cc6d6f70316e84aa9098cd72
