+++ package.json
@@ -13,13 +13,16 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.4",
+    "bootstrap": "^5.3.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "react-test-renderer": "^18.2.0",
+    "sass": "^1.63.6",
     "styled-components": "^5.3.9",
     "web-vitals": "^2.1.4"
   },
