+++ src/config/SocialOAuth..jsx
@@ -1,0 +1,2 @@
+const CLIENT_ID = process.env.REACT_APP_KAKAO_API;
+const REDIRECT_URI  =
