+++ src/index.js
@@ -5,6 +5,8 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import GlobalStyle from './assets/fonts';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import './scss/App.scss';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
