+++ src/scss/App.scss
@@ -1,0 +1,52 @@
+#top-navbar--wrap {
+  position: relative;
+  z-index: 999;
+  background: #2e2e2e;
+  .top-link {
+    color: #ff9300;
+  }
+  .basic-navbar-nav {
+    .nav-link {
+      color: #ff9300;
+    }
+    .basic-nav-dropdown {
+      .dropdown-menu {
+        background: #2e2e2e;
+      }
+      .dropdown-item {
+        color: #ff9300;
+      }
+    }
+  }
+}
+#footer-wrapper {
+  display: flex;
+  justify-content: center;
+  align-content: center;
+  align-items: center;
+  position: relative;
+  background: #2e2e2e;
+  text-align: center;
+  height: 250px;
+
+  .card-wrap {
+    background: #2e2e2e !important;
+    .card-title {
+      font-size: 1.5em;
+      color: #ff9300;
+      margin-top: 10px;
+    }
+    .cardUi {
+      li {
+        list-style-type: none;
+        color: rgb(183, 172, 172);
+      }
+      h6 {
+        display: inline;
+        margin-left: 5px;
+        margin-right: 2px;
+        color: gray;
+      }
+    }
+  }
+}
